Update so book buttons work
Note this uses the getLibrary() function which just returns a reference to the actual library. I think this is a tad unsafe and so changing this function would probably be a good idea, but for now I'm leaving it as is to get the rest of the code working.

index.js
- Make a variable hold the Library which is then used as the array to work on in the libraryEle event listener which the buttons are used with